requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.3
djangorestframework==3.12.2
idna==2.10
mysql==0.0.2
mysqlclient==2.0.1
Pillow==8.0.1
psycopg2==2.8.6
Pygments==2.7.4
PyMySQL==0.10.1
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
